pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "API"
authors = ["Rui <rui@moonsync.app>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.68.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
gunicorn = "^20.1.0"
python-dotenv = "^1.0.1"
openai = "^1.23.2"


[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
